Use signals for selection sources
