+++ src/packages/selection/Selection.test.tsx
@@ -6,8 +6,8 @@ import { createServiceOptions, setupMap } from "@open-pioneer/map-test-utils";
 import { PackageContextProvider } from "@open-pioneer/test-utils/react";
 import { act, fireEvent, render, screen, waitFor } from "@testing-library/react";
 import { Selection, SelectionCompleteEvent, SelectionSourceChangedEvent } from "./Selection";
-import { FakePointSelectionSource, NoStatusSelectionSource } from "./selectionSources";
-import { VectorLayerSelectionSourceImpl } from "./selectionSources";
+import { FakePointSelectionSource, NoStatusSelectionSource } from "./test-utils";
+import { VectorLayerSelectionSourceImpl } from "./VectorSelectionSource";
 import { NotificationService } from "@open-pioneer/notifier";
 import { Point } from "ol/geom";
 import { SelectionSource } from "./api";
