+++ src/packages/selection/package.json
@@ -17,12 +17,14 @@
         "build": "build-pioneer-package"
     },
     "peerDependencies": {
+        "@conterra/reactivity-core": "^0.4.0",
         "@chakra-ui/icons": "^2.1.1",
         "@open-pioneer/chakra-integration": "^1.1.3",
         "@open-pioneer/core": "^1.2.3",
         "@open-pioneer/map": "workspace:^",
         "@open-pioneer/notifier": "^0.3.5",
         "@open-pioneer/react-utils": "^1.0.0",
+        "@open-pioneer/reactivity": "^0.1.0",
         "@open-pioneer/runtime": "^2.1.6",
         "chakra-react-select": "^4.7.6",
         "ol": "^9.2.4",
