+++ src/packages/selection/services.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
 import { type PackageIntl, ServiceOptions } from "@open-pioneer/runtime";
-import { VectorLayerSelectionSourceImpl } from "./selectionSources";
+import { VectorLayerSelectionSourceImpl } from "./VectorSelectionSource";
 import {
     VectorLayerSelectionSource,
     VectorLayerSelectionSourceFactory,
